Moved fragment to proper package.
