+++ app/src/main/java/com/awadhesh22791/todoapp/fragment/NetbankingPaymentFragment.java
@@ -1,4 +1,4 @@
-package com.awadhesh22791.todoapp;
+package com.awadhesh22791.todoapp.fragment;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.awadhesh22791.todoapp.R;
 
 
 /**
