+++ app/src/main/res/layout/fragment_netbanking_payment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NetbankingPaymentFragment">
+    tools:context=".fragment.NetbankingPaymentFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
